refactor: split debug_ui.cc into unity-build .inc.cc files
Split the 2268-line debug_ui.cc into 8 focused files using a
unity build pattern: the main .cc file #includes all .inc.cc
fragments, so they compile as a single translation unit with
zero build system changes and identical compile times.

File structure:
- debug_ui.cc (472 lines) — core lifecycle, menu bar, dispatch
- debug_ui_helpers.inc.cc (479) — shared Lua/ImGui utilities
- debug_ui_panels.inc.cc (450) — Audio/Memory/Renderer/Camera/Physics
- debug_ui_assets.inc.cc (218) — asset viewer with tab methods
- debug_ui_watch.inc.cc (179) — watch expressions + Lua/Fennel REPL
- debug_ui_performance.inc.cc (156) — frame time graph + breakdown
- debug_ui_log.inc.cc (137) — log console with eval
- debug_ui_docs.inc.cc (128) — API documentation browser
- debug_ui_inspector.inc.cc (70) — entity inspector
